# Fetch latest trend from gems table