Ref: https://github.com/xmisao/bestgems.org/issues/32

* Improve search page rendering performance
* Improve gem page rendering performance