@xmisao 2017-10-22
Siege benchmark results.
It is 2 times faster.

### Before

Ref: https://github.com/xmisao/bestgems.org/pull/33

```
$ siege -b -c 4 -r 10 -f after_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 4 concurrent users for battle.
The server is now under siege...
Transactions:                    160 hits
Availability:                 100.00 %
Elapsed time:                  13.58 secs
Data transferred:               4.33 MB
Response time:                  0.33 secs
Transaction rate:              11.78 trans/sec
Throughput:                     0.32 MB/sec
Concurrency:                    3.89
Successful transactions:          80
Failed transactions:               0
Longest transaction:            1.31
Shortest transaction:           0.01
```

### After

```
$ siege -b -c 4 -r 10 -f bestgems_url.txt 
[alert] Zip encoding disabled; siege requires zlib support to enable it
** SIEGE 4.0.2
** Preparing 4 concurrent users for battle.
The server is now under siege...
Transactions:                    160 hits
Availability:                 100.00 %
Elapsed time:                   6.99 secs
Data transferred:               4.33 MB
Response time:                  0.17 secs
Transaction rate:              22.89 trans/sec
Throughput:                     0.62 MB/sec
Concurrency:                    3.92
Successful transactions:          80
Failed transactions:               0
Longest transaction:            0.65
Shortest transaction:           0.01
```